refactor: extract TmuxClient interface to its own file
